<refactor> 게임 내부 로직과 멤버를 초기화하는 것을 분리
