+++ src/main/kotlin/data/PathState.kt
@@ -1,7 +1,7 @@
 package data
 
-import MOVE_MARK
-import STOP_MARK
+import util.MOVE_MARK
+import util.STOP_MARK
 
 enum class PathState(val state: String) {
     MOVE(MOVE_MARK),
