--- src/main/kotlin/domain/Referee.kt
@@ -1,22 +1,0 @@
-package domain
-
-import data.Car
-
-class Referee {
-
-    fun decideWinner(cars: List<Car>): List<String> {
-        val maxScore = cars.getMaxScore()
-
-        val winners = cars.filter { car ->
-            car.getCar().second == maxScore
-        }.map { car ->
-            car.getCar().first
-        }
-
-        return winners
-    }
-
-    private fun List<Car>.getMaxScore(): Int = this.maxBy { car ->
-        car.getCar().second
-    }.getCar().second
-}
