+++ src/main/kotlin/util/Constants.kt
@@ -1,3 +1,5 @@
+package util
+
 const val CAR_NAMES_REQUEST_MSG = "경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).\n"
 const val NUMBER_OF_TRY_REQUEST_MSG = "시도할 횟수는 몇 회인가요?\n"
 
@@ -5,6 +7,7 @@ const val RESULT = "실행 결과\n"
 const val FINAL_WINNER = "최종 우승자: "
 
 const val NUMERIC_ERROR_MSG = "[ERROR] 시도 횟수는 숫자 입력으로만 받습니다."
+const val CAR_NAME_ERROR_MSG = "[ERROR] 자동차 이름은 5자를 초과할 수 없다."
 
 const val MOVE_MARK = "-"
 const val STOP_MARK = ""
