+++ src/main/kotlin/view/InputView.kt
@@ -1,7 +1,7 @@
 package view
 
-import CAR_NAMES_REQUEST_MSG
-import NUMBER_OF_TRY_REQUEST_MSG
+import util.CAR_NAMES_REQUEST_MSG
+import util.NUMBER_OF_TRY_REQUEST_MSG
 import view.View.br
 import view.View.bw
